rename command file
