+++ src/commands/list-rooms.ts
